get rid of likelihood.py
