+++ pastas/solver/__init__.py
@@ -2,6 +2,5 @@
 """solver sub-package contains the solvers and options for parameter (uncertainty) estimation for Pastas."""
 
 from .least_squares import LeastSquares, Lmfit
-from .likelihood import GaussianLikelihood, GaussianLikelihoodAr1
 from .mcmc import Emcee
-from .objective_function import misfit
+from .objective_function import GaussianLikelihood, GaussianLikelihoodAr1, misfit
