+++ tests/test_objective_functions.py
@@ -1,4 +1,4 @@
-"""Tests for the likelihood functions in pastas.solver.likelihood."""
+"""Tests for the likelihood functions in pastas.solver.objective_function."""
 
 from typing import Any
 
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from pastas.solver.likelihood import GaussianLikelihood, GaussianLikelihoodAr1
+from pastas.solver import GaussianLikelihood, GaussianLikelihoodAr1
 
 
 class TestGaussianLikelihood:
